+++ docs/src/main/scala/slinky/docs/Navbar.scala
@@ -2,7 +2,6 @@ package slinky.docs
 
 import slinky.core.FunctionalComponent
 import slinky.core.annotations.react
-import slinky.core.facade.ReactElement
 import slinky.docs.homepage.SlinkyHorizontalLogo
 import slinky.reactrouter.Link
 import slinky.web.html._