+++ tests/src/test/scala/slinky/core/FunctionalComponentTest.scala
@@ -1,12 +1,12 @@
 package slinky.core
 
-import org.scalatest.FunSuite
 import org.scalajs.dom.document
-
 import slinky.core.facade.{React, ReactElement}
 import slinky.web.ReactDOM
 
-class FunctionalComponentTest extends FunSuite {
+import org.scalatest.funsuite.AnyFunSuite
+
+class FunctionalComponentTest extends AnyFunSuite {
   test("Can render a functional component") {
     val container = document.createElement("div")
     val component = FunctionalComponent[Int](_.toString)
@@ -94,10 +94,10 @@ class FunctionalComponentTest extends FunSuite {
   test("Cannot reuse half-built functional component") {
     val component = FunctionalComponent[Int](_.toString)
     val halfBuilt = component(1)
-    val fullyBuilt: ReactElement = halfBuilt.withKey("abc")
+    halfBuilt.withKey("abc"): ReactElement
 
     assertThrows[IllegalStateException] {
-      val fullyBuilt2: ReactElement = halfBuilt.withKey("abc2")
+      halfBuilt.withKey("abc2"): ReactElement
     }
   }
 }