+++ tests/src/test/scala/slinky/core/SuspenseTest.scala
@@ -1,12 +1,13 @@
 package slinky.core
 
 import org.scalajs.dom
-import org.scalatest.FunSuite
 import slinky.core.facade.Suspense
 import slinky.web.ReactDOM
 import slinky.web.html.div
 
-class SuspenseTest extends FunSuite {
+import org.scalatest.funsuite.AnyFunSuite
+
+class SuspenseTest extends AnyFunSuite {
   test("Can render a Suspense component with children") {
     val target = dom.document.createElement("div")
     ReactDOM.render(